调整移除group相关接口后的代码

+++ weixin-java-mp/src/main/java/me/chanjar/weixin/mp/api/impl/WxMpUserTagServiceImpl.java
@@ -1,9 +1,16 @@
 package me.chanjar.weixin.mp.api.impl;
 
+import java.util.List;
+
+import org.apache.commons.lang3.StringUtils;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
 import com.google.gson.JsonArray;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonParser;
 import com.google.gson.reflect.TypeToken;
+
 import me.chanjar.weixin.common.bean.result.WxError;
 import me.chanjar.weixin.common.exception.WxErrorException;
 import me.chanjar.weixin.mp.api.WxMpService;
@@ -11,11 +18,6 @@ import me.chanjar.weixin.mp.api.WxMpUserTagService;
 import me.chanjar.weixin.mp.bean.tag.WxTagListUser;
 import me.chanjar.weixin.mp.bean.tag.WxUserTag;
 import me.chanjar.weixin.mp.util.json.WxMpGsonBuilder;
-import org.apache.commons.lang3.StringUtils;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import java.util.List;
 
 /**
  *
@@ -58,7 +60,7 @@ public class WxMpUserTagServiceImpl implements WxMpUserTagService {
   }
 
   @Override
-  public Boolean tagUpdate(Integer id, String name) throws WxErrorException {
+  public Boolean tagUpdate(Long id, String name) throws WxErrorException {
     String url = API_URL_PREFIX + "/update";
 
     JsonObject json = new JsonObject();
@@ -78,7 +80,7 @@ public class WxMpUserTagServiceImpl implements WxMpUserTagService {
   }
 
   @Override
-  public Boolean tagDelete(Integer id) throws WxErrorException {
+  public Boolean tagDelete(Long id) throws WxErrorException {
     String url = API_URL_PREFIX + "/delete";
 
     JsonObject json = new JsonObject();
@@ -98,7 +100,8 @@ public class WxMpUserTagServiceImpl implements WxMpUserTagService {
   }
 
   @Override
-  public WxTagListUser tagListUser(Integer tagId, String nextOpenid) throws WxErrorException {
+  public WxTagListUser tagListUser(Long tagId, String nextOpenid)
+      throws WxErrorException {
     String url = "https://api.weixin.qq.com/cgi-bin/user/tag/get";
 
     JsonObject json = new JsonObject();
@@ -112,7 +115,8 @@ public class WxMpUserTagServiceImpl implements WxMpUserTagService {
   }
 
   @Override
-  public boolean batchTagging(Integer tagId, String[] openids) throws WxErrorException {
+  public boolean batchTagging(Long tagId, String[] openids)
+      throws WxErrorException {
     String url = "https://api.weixin.qq.com/cgi-bin/tags/members/batchtagging";
 
     JsonObject json = new JsonObject();
@@ -135,7 +139,8 @@ public class WxMpUserTagServiceImpl implements WxMpUserTagService {
   }
 
   @Override
-  public boolean batchUntagging(Integer tagId, String[] openids) throws WxErrorException {
+  public boolean batchUntagging(Long tagId, String[] openids)
+      throws WxErrorException {
     String url = "https://api.weixin.qq.com/cgi-bin/tags/members/batchuntagging";
 
     JsonObject json = new JsonObject();
